Update response message

+++ main-service/src/service/client.service.ts
@@ -1,52 +1,73 @@
 import { prisma } from "../db/client";
 import { Client } from "@prisma/client";
 import { ClientSchema } from "../schema/client.schema";
+import axios from "axios";
+import config from "../config";
+import { ResponseService } from "../types";
+import { makeResponse } from "../utils";
 
-export async function createClient(
-  payload: ClientSchema
-): Promise<Client | null> {
-  return await prisma.client.create({
-    data: payload,
-  });
+export async function createClient(payload: ClientSchema): Promise<ResponseService<null>> {
+  try {
+    await prisma.client.create({
+      data: payload,
+    });
+    return makeResponse(201, "Client berhasil dibuat", null);
+  } catch (error) {
+    return makeResponse(400, "Client gagal dibuat", null);
+  }
 }
 
-export async function getAllClients(): Promise<Client[] | null> {
-  const clients = await prisma.client.findMany({
-    include: { project: true },
-  });
-  if (!clients.length) return null;
-  return clients;
+export async function getAllClients(): Promise<ResponseService<Client[] | null>> {
+  const clients = await prisma.client.findMany();
+
+  if (!clients.length) return makeResponse(404, "Client tidak ada", null);
+  return makeResponse(200, "Success", clients);
 }
 
-export async function getClientById(clientId: string): Promise<Client | null> {
-  return await prisma.client.findUnique({
+export async function getClientById(clientId: string): Promise<ResponseService<Client | null>> {
+  const client = await prisma.client.findUnique({
     where: { id: clientId },
     include: { project: true },
   });
+
+  if (!client) return makeResponse(404, "Client tidak ditemukan", null);
+  return makeResponse(200, "Success", client);
 }
 
 export async function updateClientById(
   clientId: string,
   payload: ClientSchema
-): Promise<Client | null> {
+): Promise<ResponseService<any>> {
   try {
-    return await prisma.client.update({
+    await prisma.client.update({
       where: { id: clientId },
       data: payload,
     });
+    return makeResponse(200, "Client berhasil diperbarui", null);
   } catch (error) {
-    return null;
+    return makeResponse(400, "Client gagal diperbarui", error);
   }
 }
 
-export async function deleteClientById(
-  clientId: string
-): Promise<Client | null> {
+export async function deleteClientById(clientId: string): Promise<ResponseService<null>> {
+  const client = await prisma.client.findUnique({
+    where: { id: clientId },
+    include: { project: true },
+  });
+
+  if (!client) return makeResponse(404, "Client tidak ditemukan", null);
+
+  const projectIds = client.project.map((p) => p.id);
+
   try {
-    return await prisma.client.delete({
-      where: { id: clientId },
-    });
+    await axios.post(`${config.api.task}/event/delete-kategori-task`, { projectIds });
   } catch (error) {
-    return null;
+    return makeResponse(500, "Terjadi kesalahan, silakan coba lagi nanti", null);
   }
+
+  await prisma.client.delete({
+    where: { id: clientId },
+  });
+
+  return makeResponse(200, "Client berhasil dihapus", null);
 }
